Revert "feat(ai-explain): add ai-explain to code blocks (#9188)"
This reverts commit e342081cbf92073ca2071e8af8a9a329b05f3d29.
