--- client/src/document/code/render-md.ts
@@ -1,19 +1,0 @@
-import { unified } from "unified";
-import remarkGfm from "remark-gfm";
-import remarkParse from "remark-parse";
-import remarkRehype from "remark-rehype";
-import rehypeSanitize from "rehype-sanitize";
-import rehypeStringify from "rehype-stringify";
-
-export function render(md: string) {
-  return unified()
-    .use(remarkParse)
-    .use(remarkGfm)
-    .use(remarkRehype, {
-      allowDangerousHtml: true,
-    })
-    .use(rehypeSanitize)
-    .use(rehypeStringify)
-    .processSync(md)
-    .toString();
-}
