+++ client/src/document/hooks.ts
@@ -2,14 +2,13 @@ import { useEffect, useRef, useState } from "react";
 import { useLocation, useParams } from "react-router-dom";
 import { useIsServer, useLocale } from "../hooks";
 import { Doc } from "../../../libs/types/document";
-import { initPlayIframe } from "../playground/utils";
-import { addExplainButton } from "./code/ai-explain";
 import {
-  addBreakoutButton,
-  getCodeAndNodesForIframe,
-  getCodeAndNodesForIframeBySampleClass,
-} from "./code/playground";
-import { addCopyToClipboardButton } from "./code/copy";
+  EditorContent,
+  initPlayIframe,
+  SESSION_KEY,
+} from "../playground/utils";
+
+const LIVE_SAMPLE_PARTS = ["html", "css", "js"];
 
 export function useDocumentURL() {
   const locale = useLocale();
@@ -19,6 +18,175 @@ export function useDocumentURL() {
   // on any URL. We can't keep that if we're going to compare the current
   // pathname with the document's mdn_url.
   return url.endsWith("/") ? url.substring(0, url.length - 1) : url;
+}
+
+const SECTION_RE = /h[1-6]/i;
+function partOfSection(heading: Element, element: Element) {
+  if (
+    SECTION_RE.test(element.tagName) &&
+    element.tagName.toLowerCase() <= heading.tagName.toLowerCase()
+  ) {
+    return false;
+  }
+  return true;
+}
+
+function sectionForHeading(heading: Element | null): Element[] {
+  const nodes: Element[] = [];
+  if (heading === null) {
+    return [];
+  }
+  if (!SECTION_RE.test(heading.tagName)) {
+    return [...heading.children];
+  }
+  let next = heading.nextElementSibling;
+  while (next && partOfSection(heading, next)) {
+    nodes.push(next);
+    if (next.nextElementSibling === null) {
+      next = next.parentElement?.nextElementSibling?.firstElementChild || null;
+    } else {
+      next = next.nextElementSibling;
+    }
+  }
+  return nodes;
+}
+
+function closestHeading(element: Element) {
+  let prev = element;
+  while (prev.parentElement?.firstElementChild) {
+    if (SECTION_RE.test(prev.parentElement.firstElementChild.tagName)) {
+      return prev.parentElement.firstElementChild;
+    }
+    prev = prev.parentElement;
+  }
+  return null;
+}
+
+function prevHeading(heading: Element) {
+  let prev = heading;
+  while (prev.parentElement?.previousElementSibling?.firstElementChild) {
+    prev = prev.parentElement.previousElementSibling.firstElementChild;
+    if (
+      SECTION_RE.test(prev.tagName) &&
+      prev.tagName.toLowerCase() < heading.tagName.toLowerCase()
+    ) {
+      return prev;
+    }
+  }
+  return null;
+}
+
+function addBreakoutButton(
+  element: Element | null,
+  id: string,
+  code: EditorContent,
+  locale: string
+) {
+  if (!element || element.querySelector(".play-button")) return;
+  const button = document.createElement("button");
+
+  button.textContent = "Play";
+
+  button.setAttribute("class", "play-button external");
+  button.type = "button";
+  button.setAttribute("data-play", id);
+  button.title = "Open in Playground";
+  element.appendChild(button);
+
+  button.addEventListener("click", (e) => {
+    sessionStorage.setItem(SESSION_KEY, JSON.stringify(code));
+    const url = new URL(window?.location.href);
+    url.pathname = `/${locale}/play`;
+    url.hash = "";
+    url.search = "";
+    window.location.href = url.href;
+  });
+}
+
+function codeForHeading(
+  heading: Element,
+  src: string
+): { code: EditorContent; nodes: Element[] } | null {
+  const section = sectionForHeading(heading);
+
+  if (!section.length) {
+    return null;
+  }
+  const code: EditorContent = {
+    css: "",
+    html: "",
+    js: "",
+    src,
+  };
+
+  const nodes: Element[] = [];
+  for (const part of LIVE_SAMPLE_PARTS) {
+    const src = section
+      .flatMap((e) => [
+        ...e?.querySelectorAll(
+          `.${part}, pre[class*="brush:${part}"], pre[class*="${part};"]`
+        ),
+      ])
+
+      .map((e) => {
+        nodes.push(e);
+        return e.textContent;
+      })
+      .join("\n");
+    if (src) {
+      code[part] += src;
+    }
+  }
+  return nodes.length ? { code, nodes } : null;
+}
+
+function getLanguage(node: Element): string | null {
+  for (const part of LIVE_SAMPLE_PARTS) {
+    if (node.classList.contains(part)) {
+      return part;
+    }
+  }
+  return null;
+}
+
+function getCodeAndNodesForIframeBySampleClass(cls: string, src: string) {
+  const code: EditorContent = {
+    css: "",
+    html: "",
+    js: "",
+    src,
+  };
+
+  let empty = true;
+  const nodes: Element[] = [];
+  [...document.getElementsByClassName(`live-sample___${cls}`)].forEach(
+    (pre) => {
+      let lang = getLanguage(pre);
+      if (lang === null) {
+        return;
+      }
+      empty = false;
+      nodes.push(pre);
+      code[lang] += pre.textContent;
+    }
+  );
+  return empty ? null : { code, nodes };
+}
+
+function getCodeAndNodesForIframe(id: string, iframe: Element, src: string) {
+  let heading = document.getElementById(id) || closestHeading(iframe);
+  if (!heading) {
+    return null;
+  }
+  let r = codeForHeading(heading, src);
+  while (r === null) {
+    heading = prevHeading(heading);
+    if (heading === null) {
+      return null;
+    }
+    r = codeForHeading(heading, src);
+  }
+  return r;
 }
 
 export function useRunSample(doc: Doc | undefined) {
@@ -68,7 +236,7 @@ export function useRunSample(doc: Doc | undefined) {
     });
   }, [doc, isServer, locale]);
 }
-export function useCopyExamplesToClipboardAndAIExplain(doc: Doc | undefined) {
+export function useCopyExamplesToClipboard(doc: Doc | undefined) {
   const location = useLocation();
   const isServer = useIsServer();
 
@@ -81,21 +249,97 @@ export function useCopyExamplesToClipboardAndAIExplain(doc: Doc | undefined) {
       return;
     }
 
+    if (!navigator.clipboard) {
+      console.log(
+        "Copy-to-clipboard disabled because your browser does not appear to support it."
+      );
+      return;
+    }
+
     document
       .querySelectorAll("div.code-example pre:not(.hidden)")
       .forEach((element) => {
         const header = element.parentElement?.querySelector(".example-header");
-        addExplainButton(header, element);
-        if (!navigator.clipboard) {
-          console.log(
-            "Copy-to-clipboard disabled because your browser does not appear to support it."
+        // No idea how a parentElement could be falsy in practice, but it can
+        // in theory and hence in TypeScript. So to having to test for it, bail
+        // early if we have to.
+        if (!header || header.querySelector(".copy-icon")) return;
+
+        const button = document.createElement("button");
+        const span = document.createElement("span");
+        const liveregion = document.createElement("span");
+
+        span.textContent = "Copy to Clipboard";
+
+        button.setAttribute("type", "button");
+        button.setAttribute("class", "icon copy-icon");
+        span.setAttribute("class", "visually-hidden");
+        liveregion.classList.add("copy-icon-message", "visually-hidden");
+        liveregion.setAttribute("role", "alert");
+
+        button.appendChild(span);
+        header.appendChild(button);
+        header.appendChild(liveregion);
+
+        button.onclick = async () => {
+          let copiedSuccessfully = true;
+          try {
+            const text = element.textContent || "";
+            await navigator.clipboard.writeText(text);
+          } catch (err) {
+            console.error(
+              "Error when trying to use navigator.clipboard.writeText()",
+              err
+            );
+            copiedSuccessfully = false;
+          }
+
+          if (copiedSuccessfully) {
+            button.classList.add("copied");
+            showCopiedMessage(header, "Copied!");
+          } else {
+            button.classList.add("failed");
+            showCopiedMessage(header, "Error trying to copy to clipboard!");
+          }
+
+          setTimeout(
+            () => {
+              hideCopiedMessage(header);
+            },
+            copiedSuccessfully ? 1000 : 3000
           );
-          return;
-        } else {
-          addCopyToClipboardButton(element, header);
-        }
+        };
       });
   }, [doc, location, isServer]);
+}
+
+function showCopiedMessage(wrapper: Element, msg: string) {
+  const element = getCopiedMessageElement(wrapper);
+  element.textContent = msg;
+  element.classList.remove("visually-hidden");
+}
+
+function hideCopiedMessage(wrapper: Element) {
+  const element = getCopiedMessageElement(wrapper);
+  element.textContent = ""; // ensure contents change, so that they are picked up by the live region
+  if (element) {
+    element.classList.add("visually-hidden");
+  }
+}
+
+function getCopiedMessageElement(wrapper: Element) {
+  const className = "copy-icon-message";
+  let element: HTMLSpanElement | null = wrapper.querySelector(
+    `span.${className}`
+  );
+  if (!element) {
+    element = document.createElement("span");
+    element.classList.add(className);
+    element.classList.add("visually-hidden");
+    element.setAttribute("role", "alert");
+    wrapper.appendChild(element);
+  }
+  return element;
 }
 
 /**
