+++ client/src/document/index.scss
@@ -541,7 +541,6 @@ pre {
 
 .code-example {
   color: var(--code-default);
-  margin-bottom: 2rem;
   position: relative;
 
   .example-header {
@@ -551,8 +550,7 @@ pre {
     border-top-left-radius: var(--elem-radius);
     border-top-right-radius: var(--elem-radius);
     display: flex;
-    flex-wrap: wrap;
-    gap: 1rem;
+    gap: 0.5rem;
     margin: 0;
     padding: 0 1rem;
 
@@ -562,33 +560,6 @@ pre {
       border-top-left-radius: 0;
       border-top-right-radius: 0;
       margin-top: 0;
-
-      &:not(:last-child) {
-        border-bottom-left-radius: 0;
-        border-bottom-right-radius: 0;
-        margin-bottom: 0;
-      }
-    }
-
-    .ai-explain-info-toggle {
-      --icon-size: 12px;
-      background-color: var(--icon-secondary);
-      cursor: pointer;
-      mask-size: contain;
-
-      &:hover,
-      &:focus {
-        background-color: var(--icon-information);
-      }
-    }
-
-    .ai-explain-info {
-      background: var(--background-information);
-      font-size: smaller;
-      margin-bottom: 0.5rem;
-      order: 23;
-      padding: 0.5rem;
-      width: 100%;
     }
 
     .language-name {
@@ -597,68 +568,11 @@ pre {
     }
   }
 
-  .ai-explain-answer {
-    background-color: var(--background-secondary);
-    border: 1px solid var(--border-primary);
-    font: var(--type-article-p);
-
-    ~ .ai-explain-answer {
-      border-top: none;
-    }
-
-    > .ai-explain-header {
-      background-color: var(--background-primary);
-      border-bottom: 1px solid var(--border-secondary);
-      margin: 0;
-      padding-top: 1rem;
-      width: 100%;
-
-      > span {
-        padding: 0.25rem 1rem;
-      }
-
-      > pre {
-        background-color: var(--background-primary);
-        margin: 0;
-        padding: 0.25rem 1rem;
-      }
-    }
-
-    > div > p {
-      margin: 0;
-      padding: 0.5rem 1rem;
-    }
-
-    .ai-explain-feedback {
-      align-items: center;
-      display: flex;
-      flex-direction: row;
-      font-size: smaller;
-      gap: 0.5rem;
-      justify-content: flex-end;
-      padding: 0.5rem;
-
-      button {
-        cursor: pointer;
-
-        &.icon-thumbs-up:hover,
-        &.icon-thumbs-up:focus {
-          background-color: var(--icon-success);
-        }
-
-        &.icon-thumbs-down:hover,
-        &.icon-thumbs-down:focus {
-          background-color: var(--icon-critical);
-        }
-      }
-    }
-  }
-
-  .play-button,
-  .ai-explain-button {
+  .play-button {
     cursor: pointer;
     font-size: smaller;
     margin: 0;
+    margin-left: auto;
     opacity: 0.4;
 
     &:hover,
@@ -666,24 +580,16 @@ pre {
       opacity: 1;
     }
 
-    &:first-child {
-      margin-left: auto;
+    &::after {
+      background-color: var(--icon-primary);
+      content: "";
+      display: inline-block;
+      height: 10px;
+      margin-left: 4px;
+      mask-image: url("../assets/icons/external.svg");
+      mask-size: cover;
+      width: 10px;
     }
-  }
-
-  .ai-explain-button.ai-explain-highlight {
-    animation: 1s ease-in 1 ai-explain-highlight-frames;
-  }
-
-  .play-button::after {
-    background-color: var(--icon-primary);
-    content: "";
-    display: inline-block;
-    height: 10px;
-    margin-left: 4px;
-    mask-image: url("../assets/icons/external.svg");
-    mask-size: cover;
-    width: 10px;
   }
 
   .copy-icon {
@@ -700,10 +606,6 @@ pre {
     &:hover,
     &:focus {
       opacity: 1;
-    }
-
-    &:first-child {
-      margin-left: auto;
     }
   }
 
@@ -842,22 +744,5 @@ kbd {
     .sidebar {
       mask-image: none;
     }
-  }
-}
-
-@keyframes ai-explain-highlight-frames {
-  0% {
-    color: var(--text-primary);
-    opacity: 0.4;
-  }
-
-  50% {
-    color: var(--category-color);
-    opacity: 1;
-  }
-
-  100% {
-    color: var(--text-primary);
-    opacity: 0.4;
   }
 }
