+++ package.json
@@ -124,7 +124,6 @@
     "read-chunk": "^4.0.3",
     "rehype-format": "^4.0.1",
     "rehype-raw": "^6.1.1",
-    "rehype-sanitize": "^5.0.1",
     "rehype-stringify": "^9.0.3",
     "remark-gfm": "^3.0.1",
     "remark-parse": "^10.0.2",
